+++ misc/build_scripts/windows/libjpeg-turbo.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-LIBJPEG_TURBO_TAG="3.1.0"
+LIBJPEG_TURBO_TAG="3.2.0"
 MT_PATH=$(ls -1 "/c/Program Files (x86)/Windows Kits/10/bin/"*"/x64/mt.exe" 2>/dev/null | sort -V | tail -n 1)
 RUN_DIR=$(pwd)
 OUTPUT_DIR="${RUN_DIR}/bin/win/jpegtran"